🎨(models) restructure navigation xapi models code structure
xAPI events models are gathered on module. To get better tracking and
management of models and their subfields, code is restructured.
